sorted out routes from server.js
